v2.8.1: Phase 4 W1.3 ship — chain-of-trust report (JSON + PDF) for site backups
Every site backup is now downloadable as a single forensic artifact bundling
its full provenance chain — backup integrity hashes, every passive verification,
every end-to-end restore drill — in two formats from the same data:

  GET /api/backup-orchestrator/chain-report/site/{id}      → JSON
  GET /api/backup-orchestrator/chain-report/site/{id}/pdf  → typst-rendered PDF

The All Backups tab now shows a Report | JSON | PDF 3-segment control on
every site row. First PDF request lazy-installs typst v0.13.0 into
/var/lib/dockpanel/typst/ (~30 MB tarball, ~30s on a fresh box, mutex-guarded
against stampede); subsequent renders are instant. Compile timeout 30s,
install timeout 90s. 503 fallback if typst install or compile fails so the
JSON path keeps working.

Site-only for v2.8.1 because only `backups.sha256_hash` / `previous_hash` /
`chain_valid` are populated today (audit migration 20260324000000). The db
+ volume backup tables don't carry hashes yet — extending chain reports
across all three kinds is a v2.8.2 follow-up that needs a hash-columns
migration plus agent changes to compute SHA-256 during db/volume backup.

Also fixed: pre-existing /api/backup-orchestrator/health 500 once any
backup row exists. SUM(BIGINT) returns NUMERIC in postgres (since aggregating
int8 can overflow); the existing query bound it to Option<i64> without an
explicit cast. Empty tables decoded NULL fine, but a real backup row started
500ing with "INT8 not compatible with NUMERIC". Cast to ::bigint in three
sites in routes/backup_orchestrator.rs::health + the rolled-up SUM in
services/backup_policy_executor.rs. Caught by the v2.8.1 fresh-VPS test
once a synthetic backup row was seeded for the chain-report PDF round-trip
— v2.8.0's 90/90 green was an empty-DB artifact.

Tests: tests/chain-report-e2e.sh new sub-suite (16 assertions: auth gate,
404 on missing backup, JSON shape, PDF magic bytes / Content-Type /
Content-Disposition, file-size sanity), wired into full-e2e.sh alongside
the tier2-pin sub-suite. Self-provisions auth via api.env JWT mint when
DOCKPANEL_TEST_PASSWORD is unset. Skips PDF assertion when
CHAIN_REPORT_SKIP_PDF=1 and reports 503 cleanly when typst install fails
so the suite still green-lights on networks that block GitHub releases.

Fresh-VPS test: 91/91 full-e2e green (89 base + 14 tier2-pin sub-suite +
16 chain-report sub-suite, with the rolled-up sub-suite assertions counted
into the 91 total) on a freshly-provisioned Vultr Ubuntu 24.04.4 box
running v2.8.1 binaries hot-swapped over a v2.8.0 install.

+++ panel/backend/src/services/backup_policy_executor.rs
@@ -80,9 +80,10 @@ async fn tick(db: &PgPool, agent: &AgentClient, jwt_secret: &str) -> Result<(), 
         execute_policy(db, agent, policy, jwt_secret).await;
     }
 
-    // Record backup storage metric for growth tracking
+    // Record backup storage metric for growth tracking. SUM(BIGINT) returns
+    // NUMERIC in postgres — cast to bigint so sqlx can decode into i64.
     let total_storage: Option<(i64,)> = sqlx::query_as(
-        "SELECT COALESCE(SUM(size_bytes), 0) FROM ( \
+        "SELECT COALESCE(SUM(size_bytes), 0)::bigint FROM ( \
             SELECT size_bytes FROM backups UNION ALL \
             SELECT size_bytes FROM database_backups UNION ALL \
             SELECT size_bytes FROM volume_backups \
